GUILD-522: Rename classes to 'Class' and 'Record' to reduce redundant names
