+++ lib/generators/deimos/schema_class/templates/Guardfile.tt
