+++ spec/schema_classes/generated.rb
@@ -1,50 +1,42 @@
-require 'deimos/schema_model'
+# frozen_string_literal: true
 
 module Deimos
   # :nodoc:
-  class GeneratedSchema < SchemaModel
-    # @!attribute [rw] status
+  class Generated < SchemaRecord
     # @return [String]
     attr_accessor :a_string
-    # @!attribute [rw] status
     # @return [Integer]
     attr_accessor :a_int
-    # @!attribute [rw] status
     # @return [Integer]
     attr_accessor :a_long
-    # @!attribute [rw] status
     # @return [Float]
     attr_accessor :a_float
-    # @!attribute [rw] status
     # @return [Float]
     attr_accessor :a_double
-    # @!attribute [rw] status
-    # @return [Deimos::GeneratedAnEnumEnum]
+    # @return [nil, Integer]
+    attr_accessor :an_optional_int
+    # @return [Deimos::AnEnum]
     attr_accessor :an_enum
-    # @!attribute [rw] status
     # @return [Array<Integer>]
     attr_accessor :an_array
-    # @!attribute [rw] status
     # @return [Hash<String, String>]
     attr_accessor :a_map
-    # @!attribute [rw] status
     # @return [String]
     attr_accessor :timestamp
-    # @!attribute [rw] status
     # @return [String]
     attr_accessor :message_id
-    # @!attribute [rw] status
-    # @return [Deimos::GeneratedARecordSchema]
+    # @return [Deimos::ARecord]
     attr_accessor :a_record
 
     # @override
-    def initialize(a_string:, a_int:, a_long:, a_float:, a_double:, an_enum:, an_array:, a_map:, timestamp:, message_id:, a_record:)
+    def initialize(a_string:, a_int:, a_long:, a_float:, a_double:, an_optional_int:, an_enum:, an_array:, a_map:, timestamp:, message_id:, a_record:)
       super()
       @a_string = a_string
       @a_int = a_int
       @a_long = a_long
       @a_float = a_float
       @a_double = a_double
+      @an_optional_int = an_optional_int
       @an_enum = an_enum
       @an_array = an_array
       @a_map = a_map
@@ -71,6 +63,7 @@ module Deimos
         'a_long' => @a_long,
         'a_float' => @a_float,
         'a_double' => @a_double,
+        'an_optional_int' => @an_optional_int,
         'an_enum' => @an_enum,
         'an_array' => @an_array,
         'a_map' => @a_map,
